Improvements based on comments on PR #1 -- Changed some filenames -- Fixed an example -- Cleaned up each.spec.js -- Added CI badge
