+++ test/data.json
@@ -1,0 +1,12 @@
+{
+  "letters": ["a", "b", "c"],
+  "people": [
+    { "name": "Pedro", "age": 19 },
+    { "name": "Juan", "age": 15 },
+    { "name": "Pablo", "age": 16 },
+    { "name": "Pancho", "age": 20 },
+    { "name": "Topo", "age": 18 },
+    { "name": "Jonathan", "age": 24 }
+  ],
+  "peopleAge": [19, 15, 16, 20, 18, 24]
+}
