--- test/testData.js
@@ -1,13 +1,0 @@
-module.exports = {
-  people: [
-    { name: 'Pedro', age: 19 },
-    { name: 'Juan', age: 15 },
-    { name: 'Pablo', age: 16 },
-    { name: 'Pancho', age: 20 },
-    { name: 'Topo', age: 18 },
-    { name: 'Jonathan', age: 24 }
-  ],
-  peopleName: ['Pedro', 'Juan', 'Pablo', 'Pancho', 'Topo', 'Jonathan'],
-  peopleAge: [19, 15, 16, 20, 18, 24],
-  peopleIdx: [0, 1, 2, 3, 4, 5]
-}
